% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_descrip.R
\name{crea_plot}
\alias{crea_plot}
\title{Function to create geoms}
\usage{
crea_plot(xval, yval, fillval, colourval, tipo)
}
\arguments{
\item{xval}{x input for the plot}

\item{yval}{y input for the plot}

\item{colourval}{variable for colouring plots}

\item{tipo}{Character the plot to be created}

\item{fillvall}{variable for filling barplots}
}
\description{
Function to create geoms
}
