% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_lm.R
\name{explain_lm}
\alias{explain_lm}
\title{Function that produces simple explanation for lm model}
\usage{
explain_lm(object)
}
\arguments{
\item{object}{An lm object}
}
\value{

}
\description{
Function that produces simple explanation for lm model
}
