% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_logistic.R
\name{explain_logistic}
\alias{explain_logistic}
\title{Function that produces simple explanation for logistic model}
\usage{
explain_logistic(object)
}
\arguments{
\item{object}{An glm-binomial object}
}
\value{

}
\description{
Function that produces simple explanation for logistic model
}
