% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_pca.R
\name{explain_pca}
\alias{explain_pca}
\title{Function that gives you insights about PCA results}
\usage{
explain_pca(x)
}
\arguments{
\item{x}{An object of class prcomp}
}
\description{
Function that gives you insights about PCA results
}
