% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_reading.R
\name{mod_data_reading_ui}
\alias{mod_data_reading_ui}
\alias{mod_data_reading_server}
\title{mod_data_reading_ui and mod_data_reading_server}
\usage{
mod_data_reading_ui(id)

mod_data_reading_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
