% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_reg_models.R
\name{mod_data_reg_models_ui}
\alias{mod_data_reg_models_ui}
\alias{mod_data_reg_models_server}
\title{mod_data_reg_models_ui and mod_data_reg_models_server}
\usage{
mod_data_reg_models_ui(id)

mod_data_reg_models_server(input, output, session, react)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
