% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_pca.R
\name{reduce_pca}
\alias{reduce_pca}
\title{Function to run PCA analysis}
\usage{
reduce_pca(x, ...)
}
\arguments{
\item{x}{A dataframe containing variables to reduce}

\item{...}{arguments to be passed to prcomp}
}
\description{
Function to run PCA analysis
}
