% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regression.R
\name{run_regression}
\alias{run_regression}
\title{Function to run regression models}
\usage{
run_regression(formula, data, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame with data}

\item{...}{other params to be passed to lm() or glm()}
}
\description{
Function to run regression models
}
