% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_clust_models.R
\name{summarised_km}
\alias{summarised_km}
\title{Runs a summarised k-means algorithm}
\usage{
summarised_km(x, scale. = TRUE, centers = 2, nstart = 25)
}
\arguments{
\item{x}{Dataframe with data to be clustered}

\item{scale.}{Whether to scale data or not}
}
\description{
Runs a summarised k-means algorithm
}
