% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_reg_models.R
\name{summarised_lm}
\alias{summarised_lm}
\title{Runs a summarised linear regression model}
\usage{
summarised_lm(formula, data, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame with data}

\item{...}{other params to be passed to lm()}
}
\description{
Runs a summarised linear regression model
}
