% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_reg_models.R
\name{summarised_logistic}
\alias{summarised_logistic}
\title{Runs a summarised logistic regression model}
\usage{
summarised_logistic(formula, data, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame with data}

\item{...}{other params to be passed to glm()}
}
\description{
Runs a summarised logistic regression model
}
