% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_reduc_models.R
\name{summarised_pca}
\alias{summarised_pca}
\title{Runs a summarised PCA analysis}
\usage{
summarised_pca(x, ...)
}
\arguments{
\item{x}{A dataframe containing variables to reduce}

\item{...}{arguments to be passed to prcomp}
}
\description{
Runs a summarised PCA analysis
}
