% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lm.R
\name{summarize_lm}
\alias{summarize_lm}
\title{Function to run and summarize a linear regression model}
\usage{
summarize_lm(formula, data, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame with data}

\item{...}{other params to be passed to lm()}
}
\description{
Function to run and summarize a linear regression model
}
