% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_logistic.R
\name{summarize_logistic}
\alias{summarize_logistic}
\title{Function to run and summarize a logistic regression model}
\usage{
summarize_logistic(formula, data, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame with data}

\item{...}{other params to be passed to glm()}
}
\description{
Function to run and summarize a logistic regression model
}
