% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{which_plot}
\alias{which_plot}
\title{Function that selects "best" possible plot}
\usage{
which_plot(data, xval, yval)
}
\arguments{
\item{data}{A dataframe}

\item{xval}{x input for the plot}

\item{yval}{y input for the plot}
}
\description{
Function that selects "best" possible plot
}
